+++ types/configs/paste-config.d.ts
@@ -1,20 +1,22 @@
+import { SanitizerConfig } from "./sanitizer-config";
+
 /**
  * Tool onPaste configuration object
  */
 export interface PasteConfig {
   /**
-   * Array of tags Tool can substitute. 
-   * 
-   * Could also contain a sanitise-config if you need to save some tag's attribute. 
-   * For example: 
+   * Array of tags Tool can substitute.
+   *
+   * Could also contain a sanitize-config if you need to save some tag's attribute.
+   * For example:
    * [
-   *   { 
+   *   {
    *     img: { src: true },
    *   }
    * ],
    * @type string[]
    */
-  tags?: (string | object)[];
+  tags?: (string | SanitizerConfig)[];
 
   /**
    * Object of string patterns Tool can substitute.